# SWAGGER_OPENAPI/datavault-broker-openapi.yaml
openapi: 3.1.0
info:
  title: DataVault Broker API
  description: |-
    API documentation for the DataVault Broker module.
    All endpoints require authentication via the X-UserID and X-Client-Key headers.
  version: 0.0.0 (4b97dc0)
  x-generated-at: 2026-03-12T10:56:37.585546
servers:
- url: http://localhost
  description: Generated server url
security:
- X-UserID-Header: []
- X-Client-Key-Header: []
...
paths:
  /permissions/role:
    put:
      tags:
      - roles-and-permissions-controller
      summary: Update an existing role
      description: Updates an existing role
      operationId: updateRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleModel"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleModel"
    post:
      tags:
      - roles-and-permissions-controller
      summary: Create a new DataVault Role
      description: Create a new DataVault Role
      operationId: createRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleModel"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleModel"
  /permissions/roleAssignment:
    put:
      tags:
      - roles-and-permissions-controller
      summary: Update an existing DataVault Role Assignment
      description: Update an existing DataVault Role Assignment
      operationId: updateRoleAssignment
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: X-Client-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignment"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
    post:
      tags:
      - roles-and-permissions-controller
      summary: Create a new DataVault Role Assignment
      description: Create a new DataVault Role Assignment
      operationId: createRoleAssignment
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: X-Client-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleAssignment"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
...
  /auth/users/isadmin:
    post:
      tags:
      - users-controller
      summary: Check if a user is an admin
      description: Checks if a user is an administrator in the DataVault system.
      operationId: isAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateUser"
        required: true
      responses:
        "200":
          description: Returns true if the user is an admin
          content:
            application/json:
              schema:
                type: boolean
                examples:
                - "true"
...
  /vaults/{vaultId}:
    get:
      tags:
      - vaults-controller
      summary: Get a specific Vault
      description: Retrieves details for a specific Vault by its ID.
      operationId: getVault
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: vaultId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultInfo"
...
  /users/{userId}:
    get:
      tags:
      - users-controller
      summary: Get a specific DataVault User
      description: Retrieves details for a specific user by their ID.
      operationId: getUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /permissions/roleAssignments/vault/{vaultId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get all role assignments for a given vault
      description: Gets all role assignments for a given vault
      operationId: getRoleAssignmentsForVault
      parameters:
      - name: vaultId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignment"
  /permissions/roleAssignments/user/{userId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get all role assignments for a given user
      description: Gets all role assignments for a given user
      operationId: getRoleAssignmentsForUser
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignment"
  /permissions/roleAssignments/school/{schoolId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get all role assignments for a given school
      description: Gets all role assignments for a given school
      operationId: getRoleAssignmentsForSchool
      parameters:
      - name: schoolId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignment"
  /permissions/roleAssignments/role/{roleId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get all role assignments for a given role
      description: Gets all role assignments for a given role
      operationId: getRoleAssignmentsForRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleAssignment"
  /permissions/roleAssignment/{assignmentId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get a specified role assignment by ID
      description: Gets a specified role assignment by its ID
      operationId: getRoleAssignment
      parameters:
      - name: assignmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleAssignment"
  /permissions/role/{roleId}:
    get:
      tags:
      - roles-and-permissions-controller
      summary: Get a specific role by ID
      description: Gets a specific role by its ID
      operationId: getRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleModel"
    delete:
      tags:
      - roles-and-permissions-controller
      summary: Delete a role by ID
      description: Deletes a role by its ID
      operationId: deleteRole
      parameters:
      - name: roleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /groups/{groupId}:
    get:
      tags:
      - groups-controller
      summary: Get a specific DataVault Group
      description: Gets details for a specific DataVault Group by its ID.
      operationId: getGroup
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
    delete:
      tags:
      - groups-controller
      summary: Delete a Group
      description: Deletes a Group from the DataVault system. Only possible if the
        group is not attached to any vaults or has no owners.
      operationId: deleteGroup
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: groupId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: boolean
...
  /permissions/roleAssignment/{roleAssignmentId}:
    delete:
      tags:
      - roles-and-permissions-controller
      summary: Delete a DataVault Role Assignment
      description: Delete a DataVault Role Assignment
      operationId: deleteRoleAssignment
      parameters:
      - name: X-UserID
        in: header
        required: true
        schema:
          type: string
      - name: X-Client-Key
        in: header
        required: true
        schema:
          type: string
      - name: roleAssignmentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    RoleModel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
          - SCHOOL
          - VAULT
          - ADMIN
        status:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionModel"
        assignedUserCount:
          type: integer
          format: int32
        idAsString:
          type: string
    RoleAssignment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          $ref: "#/components/schemas/RoleModel"
        userId:
          type: string
        schoolId:
          type: string
        vaultId:
          type: string
        pendingVaultId:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        password:
          type: string
        email:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
...
    VaultInfo:
      type: object
      description: Information about a DataVault Vault
      properties:
        id:
          type: string
          description: The unique identifier for this vault
        creationTime:
          type: string
          format: date-time
          description: The date and time when this vault was created
        policyExpiry:
          type: string
          format: date-time
          description: The date and time when the policy will expire
        policyLastChecked:
          type: string
          format: date-time
          description: The date and time when the policy check was last carried out
        name:
          type: string
          description: The name of this vault
        description:
          type: string
          description: The description of this vault
        estimate:
          type: string
          description: Estimate of vault size
          enum:
          - UNDER_100GB
          - UNDER_10TB
          - OVER_10TB
          - UNKNOWN
        billingType:
          type: string
          description: How we are billing
          enum:
          - NA
          - GRANT_FUNDING
          - BUDGET_CODE
          - SLICE
          - ORIG
          - FEEWAIVER
          - WILL_PAY
          - BUY_NEW_SLICE
          - FUNDING_NO_OR_DO_NOT_KNOW
        notes:
          type: string
          description: Notes regarding data retention
        policyID:
          type: string
          description: The policy that applies to this vault
        policyLength:
          type: string
          description: The length of the policy that applies to this vault
        groupID:
          type: string
          description: The group which is related to this vault
        userID:
          type: string
          description: The user UUN who owns this vault
        userName:
          type: string
          description: The user name who owns this vault
        datasetID:
          type: string
          description: A reference to an external metadata record that describes this
            vault
        crisID:
          type: string
          description: Another reference to an external metadata record that describes
            this vault
        datasetName:
          type: string
          description: The name of the external metadata record that describes this
            vault
        vaultSize:
          type: integer
          format: int64
          description: The size of this vault in bytes
        policyStatus:
          type: integer
          format: int32
          description: The status of the vault policy
        grantEndDate:
          type: string
          format: date
          description: Define the minimum of time the archive will be kept
        reviewDate:
          type: string
          format: date
          description: The date by which the vault should be reviewed for decision
            as to whether it should be deleted or whether there are funds available
            to support continued storage
        numberOfDeposits:
          type: integer
          format: int64
          description: Number of Deposits in a vault
        projectId:
          type: string
          description: Project Id from Pure
        sliceID:
          type: string
          description: Slice ID from erm somewhere
        authoriser:
          type: string
          description: Authoriser of the billing
        schoolOrUnit:
          type: string
          description: School / Unit to be billed
        subunit:
          type: string
          description: Subunit to be billed
        projectTitle:
          type: string
          description: Project Title (from Grant billing fieldset)
        amountToBeBilled:
          type: number
          description: Amount to be Billed
        amountBilled:
          type: number
          description: Amount Billed
        projectSize:
          type: integer
          format: int64
          description: Sum of vaults size for a projectId
        affirmed:
          type: boolean
          description: Did the user accept the various rules on the create vault intro
            page
        pureLink:
          type: boolean
          description: Did the user accept the Pure Link rule on the summary page
        confirmed:
          type: boolean
          description: Did the user confirm the pending vault yet
        contact:
          type: string
          description: Pure Contact
        ownerId:
          type: string
          description: Pending / Vault Owner ID
        ownerName:
          type: string
          description: Vault Owner Name
        vaultCreatorId:
          type: string
          description: Pending Vault Creator ID
        creators:
          type: array
          description: Data Creators
          items:
            type: string
        nominatedDataManagerIds:
          type: array
          description: Nominated Data Managers
          items:
            type: string
        depositorIds:
          type: array
          description: Depositors
          items:
            type: string
        sliceQueryChoice:
          type: string
          description: The Billing page sliceQueryChoice radio button value
          enum:
          - "YES"
          - NO_OR_DO_NOT_KNOW
          - BUY_NEW_SLICE
        fundingQueryChoice:
          type: string
          description: The Billing page fundingQueryChoice radio button value
          enum:
          - "YES"
          - FUNDING_NO_OR_DO_NOT_KNOW
        feewaiverQueryChoice:
          type: string
          description: The Billing page feewaiverQueryChoice radio button value
          enum:
          - "YES"
          - NO_OR_DO_NOT_KNOW
        paymentDetails:
          type: string
          description: The Billing payment details.
        projectSizeStr:
          type: string
        policyStatusStr:
          type: string
        dataCreators:
          type: array
          items:
            type: string
        sizeStr:
          type: string
...
    Group:
      type: object
      properties:
        id:
          type: string
          pattern: "[a-zA-Z0-9-_/ ]+"
        name:
          type: string
        enabled:
          type: boolean
        owners:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    ValidateUser:
      type: object
      properties:
        userid:
          type: string
          description: Userid
        password:
          type: string
          description: Password
...
  securitySchemes:
    X-UserID-Header:
      type: apiKey
      description: "User ID header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_USER_ID`."
      name: X-UserID
      in: header
    X-Client-Key-Header:
      type: apiKey
      description: "Client Key header, corresponds to `org.datavaultplatform.common.util.Constants.HEADER_CLIENT_KEY`."
      name: X-Client-Key
      in: header